Adding first endpoint and services and mappers

+++ Warehouse/src/main/java/com/example/warehouse/repositories/ProductInventoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductInventoryRepository extends CrudRepository<ProductInventory, Long> {
+public interface ProductInventoryRepository extends CrudRepository<ProductInventory, Integer> {
 }
