+++ Warehouse/src/main/java/com/example/warehouse/repositories/ReportRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReportRepository extends CrudRepository<Report, Long> {
+public interface ReportRepository extends CrudRepository<Report, Integer> {
 }
