+++ Warehouse/src/main/java/com/example/warehouse/repositories/TransactionProductRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TransactionProductRepository extends CrudRepository<TransactionProduct, Long> {
+public interface TransactionProductRepository extends CrudRepository<TransactionProduct, Integer> {
 }
